+++ go.mod
@@ -3,6 +3,10 @@ module github.com/caddyserver/forwardproxy
 go 1.14
 
 require (
-	github.com/caddyserver/caddy/v2 v2.0.0-rc.3
-	golang.org/x/net v0.0.0-20200324143707-d3edc9973b7e
+	cloud.google.com/go v0.54.0 // indirect
+	github.com/caddyserver/caddy/v2 v2.2.1
+	github.com/imdario/mergo v0.3.9 // indirect
+	github.com/manifoldco/promptui v0.7.0 // indirect
+	go.uber.org/zap v1.15.0
+	golang.org/x/net v0.0.0-20200707034311-ab3426394381
 )
